Refactoring of the activation functions
This draft demonstrates how the activation functions can be handled
without the need of having a global set of static objects shared via raw
pointers.

The activations still form a class hierarchy and are applied
polymorphically but the difference is that each object requiring an
activation function maintains its own copy of a particular activation
function.

The activations are "empty" classes with a single method so they are
extremely light and cheap to both store and copy. Not depending on global
objects increases safety and is generally a better approach in case
thread safety becomes a consideration point for the NAM Core library.